Set Gain to a default value if it cannot be calculated
This fixes a bug where the beat time of a peer could occasionally jump.

If a peer measures against another peer on the same machine the consecutive
input values are very similar because the error caused by network communication
- especially Wifi - is reduced. In this case it can happen that the Gain
parameter cannot be calculated because of a divisor equaling zero.

Gain defines how easily the filter will adjust to a new condition. With gain = 1
the output equals the input, with gain = 0 the input is ignored and the output
equals the last filtered value.

Clean up: Some member variables were replaced with local variables.