Assert byte order
The byte order of a multi-character integer character constant is not defined to
be the same across compilers. The key constants - as part of messages - must
have a defined byte order to assure the correct conversion from/to network byte
order of incoming/outgoing messages.
For readability we don't want to replace these constants with explicit integer
values. For now we assert at compile time if the constant has the correct byte
order to see if clients actually run into that problem.

Standard (§6.4.4.4.10): "The value of an integer character constant containing
more than one character (e.g., 'ab'), [...] is implementation-defined."