Replace enum with static const and assure fixed type width
The underlying type of enum is not guaranteed to have the same width across
different compilers. To assure this prerequisite we explicitly use an integer
type with a fixed width of 32 bit. static const is used instead of a typed enum
since its more readable and a step towards using constexpr which currently
cannot be used because of Link's VS2013 support.