Serialize microseconds as int64_t
The general overload when serializing chrono durations would have allowed
for durations, with different <Rep> sizes, to be sent across the network.
This would have lead to clients not being able to parse incoming
messages containing durations.
As we currently only serialize microseconds, this can be fixed by casting
the duration to int64_t, given it's size is 8 byte.