Bugfix: Diverging session and client start stop state
Preface: In SetSessionState(...) mClientStartStopState is updated and the new
value is passed to handleTimelineAndStartStopStateFromClient(...) which is
executed on the controller's thread. This function updates
mSessionStartStopState in case start stop sync is enabled. On changes within
the network session handleStartStopStateFromSession(...) is called on this
particular thread updating mSessionStartStopState and - if start stop sync is
enabled - mClientStartStopState.

Bug: SetSessionState(...) is called with a startStopState. Before
handleTimelineAndStartStopStateFromClient(...) is executed on the controller's
thread handleStartStopStateFromSession() is executed. This means
mSessionStartStopState is not yet updated with the value passed to the setter.
If the startStopState coming from the network is newer than the previous
mSessionStartStopState mClientStartStopState as well as mSessionStartStopState
will be updated.
Now handleTimelineAndStartStopStateFromClient() is executed with the
startStopState previously passed to the setter. In case this is newer than the
one that came from the network it will overwrite mSessionStartStopState
but not mClientStartStopState resulting in diverging startStopStates.

Fix: Conditionally update mClientStartStopState in
handleTimelineAndStartStopStateFromClient(...). This is only necessary
in case start stop sync is enabled because othewise
mClientStartStopState is not updated from the network.