Use separate grace periods for Timeline and StartStopState
If a SessionState with a changed Timeline is committed and a SessionState is
captured within the time of the grace period the returned session state will
include the committed Timeline even if a newer Timeline came in from the
network in the meantime. This is also true for the StartStopState but it has
its own grace period. This ensures for example that if a Timeline change is
committed and a SessionState captured, an incoming StartStopState change that
was received in between these calls will be reflected in the returned
SessionState.