Add advance() to MockClock
Introduce clock that will later be used to prioritize StartStopStates
within the controller.
Make the tests more readable. Using clock.micros() to set the StartStopStates'
timestamps does not change the tests' behavior. Only StartStopStates with
timestamps that are further in the future than the current StartStopState's
timestamp should be handled.