Add singleton IoContext for the esp32 platform that can be polled from main
This is a hack that allows us to drive Link in the main on the esp.
Ideally we'd like have more granular control: Only do the measurement
related stuff with high priority and the rest with a lower one. But this
is less intrusive and seems to be good enough for now.