bugfix: timeForIsPlaying
This fixes a bug where SessionState::timeForIsPlaying() of the peer
setting the state to true would not match the state of the peers
receiving the change.
To fix this the time is not mapped to beats as an intermediate
representation for the client anymore. A new ClientStartStopState not
containing a beats member is introduced.