Don't call pow() in linearRegression()
A compiler may or may not optimize pow(x, 2) to x*x, depending on the optimizer
and its settings. It's safer to manually multiply instead of depending on the
compiler's behavior.

We observed that a pow() call in the final binary causes
HostTimeFilter::sampleTimeToHostTime() to be 7x slower (104µs with pow(x, 2) and
14µs with x*x on a 1.2GHz ARM SoC).