Use only one IoContext
Dispatching between different asio::io_services can be quite expensive.
Especially on small or embedded systems. Doing all IO operations in the
same thread does not seem to add any noticeable inaccuracies or other
penalties. So we run everything on a single thread.
The lifetime of discovery::Service is now entirely managed in the IoContext
thread.