Ignore Messages From Other IPv4 Subnets
On Linux a multicast message is dispatched to all available network interfaces.
When Link receives a state message it sends out a reply messages using the
sent-from-information of the state message.
This behavior causes an issue if multiple IPv4 subnets are present: A message
is received on an interface that is not part of the subnet of that message.
Sending a reply message will fail and an UDPSendException will be thrown.
This will trigger the rebuilding of the Link network stack of that interface
and result in an endless loop.